Tools: Register tools from btw

The btw_register_tools() function equips an ellmer chat to interface with your computational environment. Chats returned by this function have access to the tools:

btw_tool_describe_data_frame: Show the data frame or table or get information about the structure of a data frame or table.

btw_tool_get_installed_packages: Displays the name and title of all installed R packages in json format.

btw_tool_get_package_help_topics: Get available help topics for an R package.

btw_tool_get_help_page: Get help page from package.

btw_tool_get_available_vignettes_in_package: Get a table of available vignettes for an R package.

btw_tool_get_vignette_from_package: Get a package vignette in plain text.

btw_tool_describe_environment: List and describe items in an environment.

Usage

btw_register_tools(chat)

Arguments

chat

An ellmer Chat object.

Value

Registers the tools with chat, updating the chat object in place. The chat input is returned invisibly.

See also

Examples

# requires an ANTHROPIC_API_KEY
if (FALSE) { # \dontrun{
ch <- ellmer::chat_claude()

btw_register_tools(ch)
} # }