Tool: Describe installed packages

Displays the name and title of all installed R packages in JSON format.

Usage

btw_tool_get_installed_packages()

Value

Returns a single string describing installed packages as JSON.

See also

Examples

cat(btw_tool_get_installed_packages())
#> ```json [
#>   {"Package":"R6","Title":"Encapsulated Classes with Reference Semantics"},
#>   {"Package":"Rcpp","Title":"Seamless R and C++ Integration"},
#>   {"Package":"S7","Title":"An Object Oriented System Meant to Become a Successor to S3 and\nS4"},
#>   {"Package":"askpass","Title":"Password Entry Utilities for R, Git, and SSH"},
#>   {"Package":"base64enc","Title":"Tools for base64 encoding"},
#>   {"Package":"brio","Title":"Basic R Input Output"},
#>   {"Package":"bslib","Title":"Custom 'Bootstrap' 'Sass' Themes for 'shiny' and 'rmarkdown'"},
#>   {"Package":"btw","Title":"Describe R Stuff to Large Language Models"},
#>   {"Package":"cachem","Title":"Cache R Objects with Automatic Pruning"},
#>   {"Package":"callr","Title":"Call R from R"},
#>   {"Package":"cli","Title":"Helpers for Developing Command Line Interfaces"},
#>   {"Package":"clipr","Title":"Read and Write from the System Clipboard"},
#>   {"Package":"coro","Title":"'Coroutines' for R"},
#>   {"Package":"crayon","Title":"Colored Terminal Output"},
#>   {"Package":"curl","Title":"A Modern and Flexible Web Client for R"},
#>   {"Package":"desc","Title":"Manipulate DESCRIPTION Files"},
#>   {"Package":"diffobj","Title":"Diffs for R Objects"},
#>   {"Package":"digest","Title":"Create Compact Hash Digests of R Objects"},
#>   {"Package":"downlit","Title":"Syntax Highlighting and Automatic Linking"},
#>   {"Package":"dplyr","Title":"A Grammar of Data Manipulation"},
#>   {"Package":"ellmer","Title":"Chat with Large Language Models"},
#>   {"Package":"evaluate","Title":"Parsing and Evaluation Tools that Provide More Details than the\nDefault"},
#>   {"Package":"fansi","Title":"ANSI Control Sequence Aware String Functions"},
#>   {"Package":"fastmap","Title":"Fast Data Structures"},
#>   {"Package":"fontawesome","Title":"Easily Work with 'Font Awesome' Icons"},
#>   {"Package":"fs","Title":"Cross-Platform File System Operations Based on 'libuv'"},
#>   {"Package":"generics","Title":"Common S3 Generics not Provided by Base R Methods Related to\nModel Fitting"},
#>   {"Package":"glue","Title":"Interpreted String Literals"},
#>   {"Package":"highr","Title":"Syntax Highlighting for R Source Code"},
#>   {"Package":"htmltools","Title":"Tools for HTML"},
#>   {"Package":"httr2","Title":"Perform HTTP Requests and Process the Responses"},
#>   {"Package":"jquerylib","Title":"Obtain 'jQuery' as an HTML Dependency Object"},
#>   {"Package":"jsonlite","Title":"A Simple and Robust JSON Parser and Generator for R"},
#>   {"Package":"knitr","Title":"A General-Purpose Package for Dynamic Report Generation in R"},
#>   {"Package":"later","Title":"Utilities for Scheduling Functions to Execute Later with Event\nLoops"},
#>   {"Package":"lifecycle","Title":"Manage the Life Cycle of your Package Functions"},
#>   {"Package":"magrittr","Title":"A Forward-Pipe Operator for R"},
#>   {"Package":"memoise","Title":"'Memoisation' of Functions"},
#>   {"Package":"mime","Title":"Map Filenames to MIME Types"},
#>   {"Package":"openssl","Title":"Toolkit for Encryption, Signatures and Certificates Based on\nOpenSSL"},
#>   {"Package":"pillar","Title":"Coloured Formatting for Columns"},
#>   {"Package":"pkgbuild","Title":"Find Tools Needed to Build R Packages"},
#>   {"Package":"pkgconfig","Title":"Private Configuration for 'R' Packages"},
#>   {"Package":"pkgdown","Title":"Make Static HTML Documentation for a Package"},
#>   {"Package":"pkgload","Title":"Simulate Package Installation and Attach"},
#>   {"Package":"praise","Title":"Praise Users"},
#>   {"Package":"prettyunits","Title":"Pretty, Human Readable Formatting of Quantities"},
#>   {"Package":"processx","Title":"Execute and Control System Processes"},
#>   {"Package":"promises","Title":"Abstractions for Promise-Based Asynchronous Programming"},
#>   {"Package":"ps","Title":"List, Query, Manipulate System Processes"},
#>   {"Package":"purrr","Title":"Functional Programming Tools"},
#>   {"Package":"ragg","Title":"Graphic Devices Based on AGG"},
#>   {"Package":"rappdirs","Title":"Application Directories: Determine Where to Save Data, Caches,\nand Logs"},
#>   {"Package":"rcmdcheck","Title":"Run 'R CMD check' from 'R' and Capture Results"},
#>   {"Package":"repr","Title":"Serializable Representations"},
#>   {"Package":"rlang","Title":"Functions for Base Types and Core R and 'Tidyverse' Features"},
#>   {"Package":"rmarkdown","Title":"Dynamic Documents for R"},
#>   {"Package":"rprojroot","Title":"Finding Files in Project Subdirectories"},
#>   {"Package":"sass","Title":"Syntactically Awesome Style Sheets ('Sass')"},
#>   {"Package":"sessioninfo","Title":"R Session Information"},
#>   {"Package":"skimr","Title":"Compact and Flexible Summaries of Data"},
#>   {"Package":"stringi","Title":"Fast and Portable Character String Processing Facilities"},
#>   {"Package":"stringr","Title":"Simple, Consistent Wrappers for Common String Operations"},
#>   {"Package":"sys","Title":"Powerful and Reliable Tools for Running System Commands in R"},
#>   {"Package":"systemfonts","Title":"System Native Font Finding"},
#>   {"Package":"testthat","Title":"Unit Testing for R"},
#>   {"Package":"textshaping","Title":"Bindings to the 'HarfBuzz' and 'Fribidi' Libraries for Text\nShaping"},
#>   {"Package":"tibble","Title":"Simple Data Frames"},
#>   {"Package":"tidyr","Title":"Tidy Messy Data"},
#>   {"Package":"tidyselect","Title":"Select from a Set of Strings"},
#>   {"Package":"tidytemplate","Title":"'pkgdown' Templates for Core 'Tidyverse' Packages"},
#>   {"Package":"tinytex","Title":"Helper Functions to Install and Maintain TeX Live, and Compile\nLaTeX Documents"},
#>   {"Package":"utf8","Title":"Unicode Text Processing"},
#>   {"Package":"vctrs","Title":"Vector Helpers"},
#>   {"Package":"waldo","Title":"Find Differences Between R Objects"},
#>   {"Package":"whisker","Title":"{{mustache}} for R, Logicless Templating"},
#>   {"Package":"withr","Title":"Run Code 'With' Temporarily Modified Global State"},
#>   {"Package":"xfun","Title":"Supporting Functions for Packages Maintained by 'Yihui Xie'"},
#>   {"Package":"xml2","Title":"Parse XML"},
#>   {"Package":"xopen","Title":"Open System Files, 'URLs', Anything"},
#>   {"Package":"yaml","Title":"Methods to Convert R Data to YAML and Back"},
#>   {"Package":"pak","Title":"Another Approach to Package Installation"},
#>   {"Package":"KernSmooth","Title":"Functions for Kernel Smoothing Supporting Wand & Jones (1995)"},
#>   {"Package":"MASS","Title":"Support Functions and Datasets for Venables and Ripley's MASS"},
#>   {"Package":"Matrix","Title":"Sparse and Dense Matrix Classes and Methods"},
#>   {"Package":"base","Title":"The R Base Package"},
#>   {"Package":"boot","Title":"Bootstrap Functions (Originally by Angelo Canty for S)"},
#>   {"Package":"class","Title":"Functions for Classification"},
#>   {"Package":"cluster","Title":"\"Finding Groups in Data\": Cluster Analysis Extended Rousseeuw et\nal."},
#>   {"Package":"codetools","Title":"Code Analysis Tools for R"},
#>   {"Package":"compiler","Title":"The R Compiler Package"},
#>   {"Package":"datasets","Title":"The R Datasets Package"},
#>   {"Package":"foreign","Title":"Read Data Stored by 'Minitab', 'S', 'SAS', 'SPSS', 'Stata',\n'Systat', 'Weka', 'dBase', ..."},
#>   {"Package":"grDevices","Title":"The R Graphics Devices and Support for Colours and Fonts"},
#>   {"Package":"graphics","Title":"The R Graphics Package"},
#>   {"Package":"grid","Title":"The Grid Graphics Package"},
#>   {"Package":"lattice","Title":"Trellis Graphics for R"},
#>   {"Package":"methods","Title":"Formal Methods and Classes"},
#>   {"Package":"mgcv","Title":"Mixed GAM Computation Vehicle with Automatic Smoothness\nEstimation"},
#>   {"Package":"nlme","Title":"Linear and Nonlinear Mixed Effects Models"},
#>   {"Package":"nnet","Title":"Feed-Forward Neural Networks and Multinomial Log-Linear Models"},
#>   {"Package":"parallel","Title":"Support for Parallel Computation in R"},
#>   {"Package":"rpart","Title":"Recursive Partitioning and Regression Trees"},
#>   {"Package":"spatial","Title":"Functions for Kriging and Point Pattern Analysis"},
#>   {"Package":"splines","Title":"Regression Spline Functions and Classes"},
#>   {"Package":"stats","Title":"The R Stats Package"},
#>   {"Package":"stats4","Title":"Statistical Functions using S4 Classes"},
#>   {"Package":"survival","Title":"Survival Analysis"},
#>   {"Package":"tcltk","Title":"Tcl/Tk Interface"},
#>   {"Package":"tools","Title":"Tools for Package Development"},
#>   {"Package":"utils","Title":"The R Utils Package"}
#> ] ```