Describe a data frame in plain text

Describe a data frame in plain text

Usage

get_data_frame(
  data_frame,
  format = c("glimpse", "print", "json"),
  dims = c(5, 100)
)

Arguments

data_frame

A single string naming a data frame or, alternatively, the data frame itself.

format

One of "glimpse", "print", or "json".

To glimpse the data column-by-column, use "glimpse". This is particularly helpful for getting a sense of data frame column names, types, and distributions, when pairings of entries in individual rows aren't particularly important.

To just print out the data frame, use print().

To get a json representation of the data, use "json". This is particularly helpful when the pairings among entries in specific rows are important to demonstrate.

dims

The number of rows and columns to show, as a numeric vector of length two. For example, the default dims = c(5, 100) shows the first 5 rows and 100 columns, whereas dims = c(Inf, Inf) would show all of the data.

Value

A character vector containing a representation of the data frame. Will error if the named data frame is not found in the environment.

See Also

register_btw_tools() registers this function as a tool for LLMs to call.

Other get_*() functions: get_data_frame(), get_environment(), get_installed_packages(), get_package_help(), get_package_vignette(), get_package_vignettes(), get_help_page()

Examples

get_data_frame(mtcars)
#> [1] "$ mpg  <dbl> 21.0, 21.0, 22.8, 21.4, 18.7\n$ cyl  <dbl> 6, 6, 4, 6, 8\n$ disp <dbl> 160, 160, 108, 258, 360\n$ hp   <dbl> 110, 110, 93, 110, 175\n$ drat <dbl> 3.90, 3.90, 3.85, 3.08, 3.15\n$ wt   <dbl> 2.620, 2.875, 2.320, 3.215, 3.440\n$ qsec <dbl> 16.46, 17.02, 18.61, 19.44, 17.02\n$ vs   <dbl> 0, 0, 1, 1, 0\n$ am   <dbl> 1, 1, 1, 0, 0\n$ gear <dbl> 4, 4, 4, 3, 3\n$ carb <dbl> 4, 4, 1, 1, 2"

get_data_frame(mtcars, format = "print")
#> [1] "    mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb\n  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>\n1  21       6   160   110  3.9   2.62  16.5     0     1     4     4\n2  21       6   160   110  3.9   2.88  17.0     0     1     4     4\n3  22.8     4   108    93  3.85  2.32  18.6     1     1     4     1\n4  21.4     6   258   110  3.08  3.22  19.4     1     0     3     1\n5  18.7     8   360   175  3.15  3.44  17.0     0     0     3     2"

get_data_frame(mtcars, format = "json")
#> [1] "[\n  {\n    \"mpg\": 21,\n    \"cyl\": 6,\n    \"disp\": 160,\n    \"hp\": 110,\n    \"drat\": 3.9,\n    \"wt\": 2.62,\n    \"qsec\": 16.46,\n    \"vs\": 0,\n    \"am\": 1,\n    \"gear\": 4,\n    \"carb\": 4,\n    \"_row\": \"Mazda RX4\"\n  },\n  {\n    \"mpg\": 21,\n    \"cyl\": 6,\n    \"disp\": 160,\n    \"hp\": 110,\n    \"drat\": 3.9,\n    \"wt\": 2.875,\n    \"qsec\": 17.02,\n    \"vs\": 0,\n    \"am\": 1,\n    \"gear\": 4,\n    \"carb\": 4,\n    \"_row\": \"Mazda RX4 Wag\"\n  },\n  {\n    \"mpg\": 22.8,\n    \"cyl\": 4,\n    \"disp\": 108,\n    \"hp\": 93,\n    \"drat\": 3.85,\n    \"wt\": 2.32,\n    \"qsec\": 18.61,\n    \"vs\": 1,\n    \"am\": 1,\n    \"gear\": 4,\n    \"carb\": 1,\n    \"_row\": \"Datsun 710\"\n  },\n  {\n    \"mpg\": 21.4,\n    \"cyl\": 6,\n    \"disp\": 258,\n    \"hp\": 110,\n    \"drat\": 3.08,\n    \"wt\": 3.215,\n    \"qsec\": 19.44,\n    \"vs\": 1,\n    \"am\": 0,\n    \"gear\": 3,\n    \"carb\": 1,\n    \"_row\": \"Hornet 4 Drive\"\n  },\n  {\n    \"mpg\": 18.7,\n    \"cyl\": 8,\n    \"disp\": 360,\n    \"hp\": 175,\n    \"drat\": 3.15,\n    \"wt\": 3.44,\n    \"qsec\": 17.02,\n    \"vs\": 0,\n    \"am\": 0,\n    \"gear\": 3,\n    \"carb\": 2,\n    \"_row\": \"Hornet Sportabout\"\n  }\n] "