Describe installed packages

Displays the name and title of all installed R packages in json format.

Usage

get_installed_packages()

Value

A single string describing installed packages in json.

See Also

register_btw_tools() registers this function as a tool for LLMs to call.

Other get_*() functions: get_data_frame(), get_environment(), get_installed_packages(), get_package_help(), get_package_vignette(), get_package_vignettes(), get_help_page()

Examples

cat(get_installed_packages())
#> [
#>   {
#>     "Package": "R6",
#>     "Title": "Encapsulated Classes with Reference Semantics",
#>     "_row": "R6"
#>   },
#>   {
#>     "Package": "Rcpp",
#>     "Title": "Seamless R and C++ Integration",
#>     "_row": "Rcpp"
#>   },
#>   {
#>     "Package": "S7",
#>     "Title": "An Object Oriented System Meant to Become a Successor to S3 and\nS4",
#>     "_row": "S7"
#>   },
#>   {
#>     "Package": "askpass",
#>     "Title": "Password Entry Utilities for R, Git, and SSH",
#>     "_row": "askpass"
#>   },
#>   {
#>     "Package": "base64enc",
#>     "Title": "Tools for base64 encoding",
#>     "_row": "base64enc"
#>   },
#>   {
#>     "Package": "brio",
#>     "Title": "Basic R Input Output",
#>     "_row": "brio"
#>   },
#>   {
#>     "Package": "bslib",
#>     "Title": "Custom 'Bootstrap' 'Sass' Themes for 'shiny' and 'rmarkdown'",
#>     "_row": "bslib"
#>   },
#>   {
#>     "Package": "btw",
#>     "Title": "Describe R Stuff to Large Language Models",
#>     "_row": "btw"
#>   },
#>   {
#>     "Package": "cachem",
#>     "Title": "Cache R Objects with Automatic Pruning",
#>     "_row": "cachem"
#>   },
#>   {
#>     "Package": "callr",
#>     "Title": "Call R from R",
#>     "_row": "callr"
#>   },
#>   {
#>     "Package": "cli",
#>     "Title": "Helpers for Developing Command Line Interfaces",
#>     "_row": "cli"
#>   },
#>   {
#>     "Package": "clipr",
#>     "Title": "Read and Write from the System Clipboard",
#>     "_row": "clipr"
#>   },
#>   {
#>     "Package": "coro",
#>     "Title": "'Coroutines' for R",
#>     "_row": "coro"
#>   },
#>   {
#>     "Package": "crayon",
#>     "Title": "Colored Terminal Output",
#>     "_row": "crayon"
#>   },
#>   {
#>     "Package": "curl",
#>     "Title": "A Modern and Flexible Web Client for R",
#>     "_row": "curl"
#>   },
#>   {
#>     "Package": "desc",
#>     "Title": "Manipulate DESCRIPTION Files",
#>     "_row": "desc"
#>   },
#>   {
#>     "Package": "diffobj",
#>     "Title": "Diffs for R Objects",
#>     "_row": "diffobj"
#>   },
#>   {
#>     "Package": "digest",
#>     "Title": "Create Compact Hash Digests of R Objects",
#>     "_row": "digest"
#>   },
#>   {
#>     "Package": "downlit",
#>     "Title": "Syntax Highlighting and Automatic Linking",
#>     "_row": "downlit"
#>   },
#>   {
#>     "Package": "dplyr",
#>     "Title": "A Grammar of Data Manipulation",
#>     "_row": "dplyr"
#>   },
#>   {
#>     "Package": "ellmer",
#>     "Title": "Chat with Large Language Models",
#>     "_row": "ellmer"
#>   },
#>   {
#>     "Package": "evaluate",
#>     "Title": "Parsing and Evaluation Tools that Provide More Details than the\nDefault",
#>     "_row": "evaluate"
#>   },
#>   {
#>     "Package": "fansi",
#>     "Title": "ANSI Control Sequence Aware String Functions",
#>     "_row": "fansi"
#>   },
#>   {
#>     "Package": "fastmap",
#>     "Title": "Fast Data Structures",
#>     "_row": "fastmap"
#>   },
#>   {
#>     "Package": "fontawesome",
#>     "Title": "Easily Work with 'Font Awesome' Icons",
#>     "_row": "fontawesome"
#>   },
#>   {
#>     "Package": "fs",
#>     "Title": "Cross-Platform File System Operations Based on 'libuv'",
#>     "_row": "fs"
#>   },
#>   {
#>     "Package": "generics",
#>     "Title": "Common S3 Generics not Provided by Base R Methods Related to\nModel Fitting",
#>     "_row": "generics"
#>   },
#>   {
#>     "Package": "glue",
#>     "Title": "Interpreted String Literals",
#>     "_row": "glue"
#>   },
#>   {
#>     "Package": "highr",
#>     "Title": "Syntax Highlighting for R Source Code",
#>     "_row": "highr"
#>   },
#>   {
#>     "Package": "htmltools",
#>     "Title": "Tools for HTML",
#>     "_row": "htmltools"
#>   },
#>   {
#>     "Package": "httr2",
#>     "Title": "Perform HTTP Requests and Process the Responses",
#>     "_row": "httr2"
#>   },
#>   {
#>     "Package": "jquerylib",
#>     "Title": "Obtain 'jQuery' as an HTML Dependency Object",
#>     "_row": "jquerylib"
#>   },
#>   {
#>     "Package": "jsonlite",
#>     "Title": "A Simple and Robust JSON Parser and Generator for R",
#>     "_row": "jsonlite"
#>   },
#>   {
#>     "Package": "knitr",
#>     "Title": "A General-Purpose Package for Dynamic Report Generation in R",
#>     "_row": "knitr"
#>   },
#>   {
#>     "Package": "later",
#>     "Title": "Utilities for Scheduling Functions to Execute Later with Event\nLoops",
#>     "_row": "later"
#>   },
#>   {
#>     "Package": "lifecycle",
#>     "Title": "Manage the Life Cycle of your Package Functions",
#>     "_row": "lifecycle"
#>   },
#>   {
#>     "Package": "magrittr",
#>     "Title": "A Forward-Pipe Operator for R",
#>     "_row": "magrittr"
#>   },
#>   {
#>     "Package": "memoise",
#>     "Title": "'Memoisation' of Functions",
#>     "_row": "memoise"
#>   },
#>   {
#>     "Package": "mime",
#>     "Title": "Map Filenames to MIME Types",
#>     "_row": "mime"
#>   },
#>   {
#>     "Package": "openssl",
#>     "Title": "Toolkit for Encryption, Signatures and Certificates Based on\nOpenSSL",
#>     "_row": "openssl"
#>   },
#>   {
#>     "Package": "pillar",
#>     "Title": "Coloured Formatting for Columns",
#>     "_row": "pillar"
#>   },
#>   {
#>     "Package": "pkgbuild",
#>     "Title": "Find Tools Needed to Build R Packages",
#>     "_row": "pkgbuild"
#>   },
#>   {
#>     "Package": "pkgconfig",
#>     "Title": "Private Configuration for 'R' Packages",
#>     "_row": "pkgconfig"
#>   },
#>   {
#>     "Package": "pkgdown",
#>     "Title": "Make Static HTML Documentation for a Package",
#>     "_row": "pkgdown"
#>   },
#>   {
#>     "Package": "pkgload",
#>     "Title": "Simulate Package Installation and Attach",
#>     "_row": "pkgload"
#>   },
#>   {
#>     "Package": "praise",
#>     "Title": "Praise Users",
#>     "_row": "praise"
#>   },
#>   {
#>     "Package": "prettyunits",
#>     "Title": "Pretty, Human Readable Formatting of Quantities",
#>     "_row": "prettyunits"
#>   },
#>   {
#>     "Package": "processx",
#>     "Title": "Execute and Control System Processes",
#>     "_row": "processx"
#>   },
#>   {
#>     "Package": "promises",
#>     "Title": "Abstractions for Promise-Based Asynchronous Programming",
#>     "_row": "promises"
#>   },
#>   {
#>     "Package": "ps",
#>     "Title": "List, Query, Manipulate System Processes",
#>     "_row": "ps"
#>   },
#>   {
#>     "Package": "purrr",
#>     "Title": "Functional Programming Tools",
#>     "_row": "purrr"
#>   },
#>   {
#>     "Package": "ragg",
#>     "Title": "Graphic Devices Based on AGG",
#>     "_row": "ragg"
#>   },
#>   {
#>     "Package": "rappdirs",
#>     "Title": "Application Directories: Determine Where to Save Data, Caches,\nand Logs",
#>     "_row": "rappdirs"
#>   },
#>   {
#>     "Package": "rcmdcheck",
#>     "Title": "Run 'R CMD check' from 'R' and Capture Results",
#>     "_row": "rcmdcheck"
#>   },
#>   {
#>     "Package": "rlang",
#>     "Title": "Functions for Base Types and Core R and 'Tidyverse' Features",
#>     "_row": "rlang"
#>   },
#>   {
#>     "Package": "rmarkdown",
#>     "Title": "Dynamic Documents for R",
#>     "_row": "rmarkdown"
#>   },
#>   {
#>     "Package": "rprojroot",
#>     "Title": "Finding Files in Project Subdirectories",
#>     "_row": "rprojroot"
#>   },
#>   {
#>     "Package": "sass",
#>     "Title": "Syntactically Awesome Style Sheets ('Sass')",
#>     "_row": "sass"
#>   },
#>   {
#>     "Package": "sessioninfo",
#>     "Title": "R Session Information",
#>     "_row": "sessioninfo"
#>   },
#>   {
#>     "Package": "stringi",
#>     "Title": "Fast and Portable Character String Processing Facilities",
#>     "_row": "stringi"
#>   },
#>   {
#>     "Package": "sys",
#>     "Title": "Powerful and Reliable Tools for Running System Commands in R",
#>     "_row": "sys"
#>   },
#>   {
#>     "Package": "systemfonts",
#>     "Title": "System Native Font Finding",
#>     "_row": "systemfonts"
#>   },
#>   {
#>     "Package": "testthat",
#>     "Title": "Unit Testing for R",
#>     "_row": "testthat"
#>   },
#>   {
#>     "Package": "textshaping",
#>     "Title": "Bindings to the 'HarfBuzz' and 'Fribidi' Libraries for Text\nShaping",
#>     "_row": "textshaping"
#>   },
#>   {
#>     "Package": "tibble",
#>     "Title": "Simple Data Frames",
#>     "_row": "tibble"
#>   },
#>   {
#>     "Package": "tidyselect",
#>     "Title": "Select from a Set of Strings",
#>     "_row": "tidyselect"
#>   },
#>   {
#>     "Package": "tidytemplate",
#>     "Title": "'pkgdown' Templates for Core 'Tidyverse' Packages",
#>     "_row": "tidytemplate"
#>   },
#>   {
#>     "Package": "tinytex",
#>     "Title": "Helper Functions to Install and Maintain TeX Live, and Compile\nLaTeX Documents",
#>     "_row": "tinytex"
#>   },
#>   {
#>     "Package": "utf8",
#>     "Title": "Unicode Text Processing",
#>     "_row": "utf8"
#>   },
#>   {
#>     "Package": "vctrs",
#>     "Title": "Vector Helpers",
#>     "_row": "vctrs"
#>   },
#>   {
#>     "Package": "waldo",
#>     "Title": "Find Differences Between R Objects",
#>     "_row": "waldo"
#>   },
#>   {
#>     "Package": "whisker",
#>     "Title": "{{mustache}} for R, Logicless Templating",
#>     "_row": "whisker"
#>   },
#>   {
#>     "Package": "withr",
#>     "Title": "Run Code 'With' Temporarily Modified Global State",
#>     "_row": "withr"
#>   },
#>   {
#>     "Package": "xfun",
#>     "Title": "Supporting Functions for Packages Maintained by 'Yihui Xie'",
#>     "_row": "xfun"
#>   },
#>   {
#>     "Package": "xml2",
#>     "Title": "Parse XML",
#>     "_row": "xml2"
#>   },
#>   {
#>     "Package": "xopen",
#>     "Title": "Open System Files, 'URLs', Anything",
#>     "_row": "xopen"
#>   },
#>   {
#>     "Package": "yaml",
#>     "Title": "Methods to Convert R Data to YAML and Back",
#>     "_row": "yaml"
#>   },
#>   {
#>     "Package": "pak",
#>     "Title": "Another Approach to Package Installation",
#>     "_row": "pak"
#>   },
#>   {
#>     "Package": "KernSmooth",
#>     "Title": "Functions for Kernel Smoothing Supporting Wand & Jones (1995)",
#>     "_row": "KernSmooth"
#>   },
#>   {
#>     "Package": "MASS",
#>     "Title": "Support Functions and Datasets for Venables and Ripley's MASS",
#>     "_row": "MASS"
#>   },
#>   {
#>     "Package": "Matrix",
#>     "Title": "Sparse and Dense Matrix Classes and Methods",
#>     "_row": "Matrix"
#>   },
#>   {
#>     "Package": "base",
#>     "Title": "The R Base Package",
#>     "_row": "base"
#>   },
#>   {
#>     "Package": "boot",
#>     "Title": "Bootstrap Functions (Originally by Angelo Canty for S)",
#>     "_row": "boot"
#>   },
#>   {
#>     "Package": "class",
#>     "Title": "Functions for Classification",
#>     "_row": "class"
#>   },
#>   {
#>     "Package": "cluster",
#>     "Title": "\"Finding Groups in Data\": Cluster Analysis Extended Rousseeuw et\nal.",
#>     "_row": "cluster"
#>   },
#>   {
#>     "Package": "codetools",
#>     "Title": "Code Analysis Tools for R",
#>     "_row": "codetools"
#>   },
#>   {
#>     "Package": "compiler",
#>     "Title": "The R Compiler Package",
#>     "_row": "compiler"
#>   },
#>   {
#>     "Package": "datasets",
#>     "Title": "The R Datasets Package",
#>     "_row": "datasets"
#>   },
#>   {
#>     "Package": "foreign",
#>     "Title": "Read Data Stored by 'Minitab', 'S', 'SAS', 'SPSS', 'Stata',\n'Systat', 'Weka', 'dBase', ...",
#>     "_row": "foreign"
#>   },
#>   {
#>     "Package": "grDevices",
#>     "Title": "The R Graphics Devices and Support for Colours and Fonts",
#>     "_row": "grDevices"
#>   },
#>   {
#>     "Package": "graphics",
#>     "Title": "The R Graphics Package",
#>     "_row": "graphics"
#>   },
#>   {
#>     "Package": "grid",
#>     "Title": "The Grid Graphics Package",
#>     "_row": "grid"
#>   },
#>   {
#>     "Package": "lattice",
#>     "Title": "Trellis Graphics for R",
#>     "_row": "lattice"
#>   },
#>   {
#>     "Package": "methods",
#>     "Title": "Formal Methods and Classes",
#>     "_row": "methods"
#>   },
#>   {
#>     "Package": "mgcv",
#>     "Title": "Mixed GAM Computation Vehicle with Automatic Smoothness\nEstimation",
#>     "_row": "mgcv"
#>   },
#>   {
#>     "Package": "nlme",
#>     "Title": "Linear and Nonlinear Mixed Effects Models",
#>     "_row": "nlme"
#>   },
#>   {
#>     "Package": "nnet",
#>     "Title": "Feed-Forward Neural Networks and Multinomial Log-Linear Models",
#>     "_row": "nnet"
#>   },
#>   {
#>     "Package": "parallel",
#>     "Title": "Support for Parallel Computation in R",
#>     "_row": "parallel"
#>   },
#>   {
#>     "Package": "rpart",
#>     "Title": "Recursive Partitioning and Regression Trees",
#>     "_row": "rpart"
#>   },
#>   {
#>     "Package": "spatial",
#>     "Title": "Functions for Kriging and Point Pattern Analysis",
#>     "_row": "spatial"
#>   },
#>   {
#>     "Package": "splines",
#>     "Title": "Regression Spline Functions and Classes",
#>     "_row": "splines"
#>   },
#>   {
#>     "Package": "stats",
#>     "Title": "The R Stats Package",
#>     "_row": "stats"
#>   },
#>   {
#>     "Package": "stats4",
#>     "Title": "Statistical Functions using S4 Classes",
#>     "_row": "stats4"
#>   },
#>   {
#>     "Package": "survival",
#>     "Title": "Survival Analysis",
#>     "_row": "survival"
#>   },
#>   {
#>     "Package": "tcltk",
#>     "Title": "Tcl/Tk Interface",
#>     "_row": "tcltk"
#>   },
#>   {
#>     "Package": "tools",
#>     "Title": "Tools for Package Development",
#>     "_row": "tools"
#>   },
#>   {
#>     "Package": "utils",
#>     "Title": "The R Utils Package",
#>     "_row": "utils"
#>   }
#> ]