Describe R package documentation

These functions describe package documentation in plain text:

Usage

get_package_help(package_name)

get_help_page(package_name, topic)

Arguments

package_name

The name of the package as a string, e.g. "shiny".

topic

The topic_id or alias of the help page, e.g. "withProgress" or "incProgress". Find topic_ids or aliases using get_package_help().

Value

get_package_help() returns the topic_id, title, and aliases fields for every topic in a package's documentation as a json-formatted string.

get_help_page() return the help-page for a package topic as a string.

See Also

register_btw_tools() registers this function as a tool for LLMs to call.

Other get_*() functions: get_data_frame(), get_environment(), get_installed_packages(), get_package_help(), get_help_page()

Examples

cat(get_package_help("btw"))
#> [
#>   {
#>     "topic_id": "btw",
#>     "title": "Plain-text descriptions of R objects",
#>     "aliases": ["btw"]
#>   },
#>   {
#>     "topic_id": "btw-package",
#>     "title": "btw: Describe R Stuff in Plain Text",
#>     "aliases": ["btw-package"]
#>   },
#>   {
#>     "topic_id": "get_data_frame",
#>     "title": "Describe a data frame in plain text",
#>     "aliases": ["get_data_frame"]
#>   },
#>   {
#>     "topic_id": "get_environment",
#>     "title": "Describe the contents of an environment",
#>     "aliases": ["get_environment"]
#>   },
#>   {
#>     "topic_id": "get_installed_packages",
#>     "title": "Describe installed packages",
#>     "aliases": ["get_installed_packages"]
#>   },
#>   {
#>     "topic_id": "get_pkg",
#>     "title": "Describe R package documentation",
#>     "aliases": ["get_pkg", "get_package_help", "get_help_page"]
#>   },
#>   {
#>     "topic_id": "register_btw_tools",
#>     "title": "Register tools from btw",
#>     "aliases": ["register_btw_tools"]
#>   }
#> ] 

cat(get_help_page("btw", "btw"))
#> btw                    package:btw                     R Documentation
#> 
#> Plain-text descriptions of R objects
#> 
#> Description:
#> 
#>      This function allows you to quickly describe your computational
#>      environment to a model by concatenating plain-text descriptions of
#>      "R stuff", from data frames to packages to function documentation.
#> 
#> Usage:
#> 
#>      btw(..., clipboard = interactive())
#>      
#> Arguments:
#> 
#>      ...: Objects to describe from your R environment. You can pass
#>           objects themselves, like data frames or functions, or the
#>           function also accepts output from get_*() functions like
#>           ‘get_package_help()’, ‘get_help_page()’, etc. If omitted,
#>           this function will just describe the elements in your global
#>           R environment.
#> 
#> clipboard: Whether to write the results to the clipboard. A single
#>           logical value; will default to ‘TRUE’ when run interactively.
#> 
#> Value:
#> 
#>      The combined elements as a string, invisibly. If ‘clipboard’ is
#>      ‘TRUE’, the result is also written to the system clipboard.
#> 
#> Examples:
#> 
#>      btw()
#>      
#>      btw(mtcars)
#>      
#>      btw(btw::btw)
#>