Register tools from btw

The register_btw_tools() function equips an ellmer chat to interface with your computational environment. Chats returned by this function have access to the tools get_data_frame(), get_environment(), get_installed_packages(), get_package_help(), get_package_vignette(), get_package_vignettes(), get_help_page().

Usage

register_btw_tools(chat)

Arguments

chat

An ellmer Chat object.

Value

The chat object with tools registered.

Examples

# requires an ANTHROPIC_API_KEY
if (FALSE) { # \dontrun{
ch <- ellmer::chat_claude()

register_btw_tools(ch)
} # }