Plain-text descriptions of R objects

This function allows you to quickly describe your computational environment to a model by concatenating plain-text descriptions of "R stuff", from data frames to packages to function documentation.

Usage

btw(..., clipboard = interactive())

Arguments

...

Objects to describe from your R environment. You can pass objects themselves, like data frames or functions, or the function also accepts output from get_*() functions like get_package_help(), get_help_page(), etc. If omitted, this function will just describe the elements in your global R environment.

clipboard

Whether to write the results to the clipboard. A single logical value; will default to TRUE when run interactively.

Value

The combined elements as a string, invisibly. If clipboard is TRUE, the result is also written to the system clipboard.

Examples

btw()

btw(mtcars)

btw(btw::btw)