Create a btw-enhanced ellmer chat client

Creates an ellmer::Chat client, enhanced with the tools from btw_register_tools(). Use btw_chat() to create the chat client for general or interactive use at the console, or btw_app() to create a chat client and launch a Shiny app for chatting with a btw-enhanced LLM in your local workspace.

Usage

btw_client(..., client = NULL, include = NULL)

btw_app(..., client = NULL, include = NULL)

Arguments

...

Objects and documentation to be included as context in the chat, via btw().

client

An ellmer::Chat client, defaults to ellmer::chat_claude(). You can use the btw.chat_client option to set a default client for new btw_client() calls.

include

Names of tools or tool groups to include when registering tools, e.g. include = "docs" to include only the documentation related tools, or include = c("data", "docs", "environment"), etc.

Value

Returns an ellmer::Chat object with additional tools registered by btw_register_tools(). btw_app() returns the chat object invisibly, and the chat object with the messages added during the chat session.

Details

Client Options

btw.chat_client: The ellmer::Chat client to use as the basis for new btw_client() or btw_app() chats.

btw.chat_include: The btw tools to include by default when starting a new btw chat, see btw_register_tools() for details.

Functions

btw_client(): Create a btw-enhanced ellmer::Chat client

btw_app(): Create a btw-enhanced client and launch a Shiny app to chat

Examples

if (interactive()) {
  withr::local_options(list(
    btw.chat_client = ellmer::chat_ollama(model="llama3.1:8b")
  ))

  btw_chat <- btw_client()
  chat$chat("How can I replace `stop()` calls with functions from the cli package?")
}