Tools: Register tools from btw

The btw_register_tools() function equips an ellmer chat to interface with your computational environment. Chats returned by this function have access to the tools:

Name	Group	Description

btw_tool_describe_data_frame	data	Show the data frame or table or get information about the structure of a data frame or table.

btw_tool_get_available_vignettes_in_package	docs	List available vignettes for an R package.

btw_tool_get_help_page	docs	Get help page from package.

btw_tool_get_installed_packages	docs	Displays the name and title of all installed R packages in json format.

btw_tool_get_package_help_topics	docs	Get available help topics for an R package.

btw_tool_get_vignette_from_package	docs	Get a package vignette in plain text.

btw_tool_describe_environment	environment	List and describe items in an environment.

btw_tool_list_files	files	List files in the current working directory or in subfolders in the current project directory.

btw_tool_read_text_file	files	Read an entire text file.

btw_tool_read_current_editor	ide	Read the contents of the editor that is currently open in the user's IDE.

Usage

btw_register_tools(chat, include = NULL)

Arguments

chat

An ellmer Chat object.

include

Names of tools or tool groups to include when registering tools, e.g. include = "docs" to include only the documentation related tools, or include = c("data", "docs", "environment"), etc.

Value

Registers the tools with chat, updating the chat object in place. The chat input is returned invisibly.

See also

Examples

# requires an ANTHROPIC_API_KEY
if (FALSE) { # \dontrun{
ch <- ellmer::chat_claude()

btw_register_tools(ch)
} # }