Tool: Describe user's platform

Describes the R version, operating system, and language and locale settings for the user's system. When using btw_client() or btw_app(), this information is automatically included in the system prompt.

Usage

btw_tool_session_platform_info()

Value

Returns a string describing the user's platform.

See also

Examples

cat(btw_tool_session_platform_info())
#> <system_info>
#> R_VERSION: R version 4.5.0 (2025-04-11)
#> OS: Ubuntu 24.04.2 LTS
#> SYSTEM: x86_64, linux-gnu
#> UI: X11
#> LANGUAGE: en
#> LOCALE: C
#> ENCODING: C.UTF-8
#> TIMEZONE: UTC
#> DATE: Monday, April 21, 2025 (2025-04-21)
#> </system_info>