Tool: Describe user's platform

Describes the R version, operating system, and language and locale settings for the user's system. When using btw_client() or btw_app(), this information is automatically included in the system prompt.

Usage

btw_tool_session_platform_info(`_intent` = "")

Arguments

_intent

An optional string describing the intent of the tool use. When the tool is used by an LLM, the model will use this argument to explain why it called the tool.

Value

Returns a string describing the user's platform.

See also

Examples

btw_tool_session_platform_info()
#> <btw::BtwSessionInfoToolResult>
#>  @ value  : chr "<system_info>\nR_VERSION: R version 4.5.1 (2025-06-13)\nOS: Ubuntu 24.04.2 LTS\nSYSTEM: x86_64, linux-gnu\nUI: "| __truncated__
#>  @ error  : NULL
#>  @ extra  :List of 9
#>  .. $ r_version:: chr "R version 4.5.1 (2025-06-13)"
#>  .. $ os:       : chr "Ubuntu 24.04.2 LTS"
#>  .. $ system:   : chr "x86_64, linux-gnu"
#>  .. $ ui:       : chr "X11"
#>  .. $ language: : chr "en"
#>  .. $ locale:   : chr "C"
#>  .. $ encoding: : chr "C.UTF-8"
#>  .. $ timezone: : chr "UTC"
#>  .. $ date:     : chr "Friday, August 22, 2025 (2025-08-22)"
#>  .. - attr(*, "class")= chr [1:2] "platform_info" "list"
#>  @ request: NULL